security+updating: Rewrote updating.
This rewrites and fixes issues with updating,
updates now happen in main process and updates are
checked in preload before the user even sees the
launch button, so launching should take slightly
less time to happen.

This is part of the PR (security/remove-node-integration)
that removes nodeIntegration usage and replaces it with
preload scripts.

Signed-off-by: TheOddGarlic <umutinanerdogan62@gmail.com>